+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/estate/entity/Estate.java
@@ -1,4 +1,4 @@
-package com.piehouse.wooreadmin.dashboard.entity;
+package com.piehouse.wooreadmin.estate.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
