+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/estate/repository/EstateRepository.java
@@ -1,7 +1,7 @@
-package com.piehouse.wooreadmin.dashboard.repository;
+package com.piehouse.wooreadmin.estate.repository;
 
-import com.piehouse.wooreadmin.dashboard.entity.Estate;
-import com.piehouse.wooreadmin.dashboard.entity.SubState;
+import com.piehouse.wooreadmin.estate.entity.Estate;
+import com.piehouse.wooreadmin.estate.entity.SubState;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,7 +11,7 @@ import java.util.List;
 public interface EstateRepository extends JpaRepository<Estate, Long> {
 
     @Query("SELECT e FROM Estate e JOIN FETCH e.agent WHERE e.subState = :state")
-    List<Estate> findWithAgentById(@Param("state")SubState state);
+    List<Estate> findWithAgentById(@Param("state") SubState state);
 
     List<Estate> findEstateBySubState(SubState state);
 }
