+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/estate/service/EstateService.java
@@ -1,0 +1,16 @@
+package com.piehouse.wooreadmin.estate.service;
+
+import com.piehouse.wooreadmin.estate.dto.EstateApproveRequest;
+import com.piehouse.wooreadmin.estate.entity.Estate;
+
+import java.util.List;
+
+public interface EstateService {
+
+    List<Estate> getAllEstate();
+
+    Boolean approveEstate(EstateApproveRequest dto);
+
+    Boolean rejectEstate(EstateApproveRequest dto);
+
+}
