+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/subscription/entity/Subscription.java
@@ -1,6 +1,6 @@
 package com.piehouse.wooreadmin.subscription.entity;
 
-import com.piehouse.wooreadmin.dashboard.entity.Estate;
+import com.piehouse.wooreadmin.estate.entity.Estate;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
